refact: PlaceOrderScreen removes duplicates, extracts similar OrderSumCard
